refactor(dashboard): hand-write executeMigration client instead of orval operation override
