+++ dashboard/src/features/orgs/projects/database/dataGrid/hooks/useComputedFieldMetadataMutation/deleteComputedFieldMigration.ts
@@ -1,4 +1,4 @@
-import { executeMigration } from '@/utils/hasura-api/generated/default/default';
+import { executeMigration } from '@/utils/hasura-api/migrationFetch';
 import type {
   ComputedFieldItem,
   DropComputedFieldArgs,
