+++ dashboard/src/features/orgs/projects/database/dataGrid/hooks/useComputedFieldMetadataMutation/editComputedFieldMigration.ts
@@ -1,4 +1,4 @@
-import { executeMigration } from '@/utils/hasura-api/generated/default/default';
+import { executeMigration } from '@/utils/hasura-api/migrationFetch';
 import type {
   AddComputedFieldArgs,
   ComputedFieldItem,
