+++ dashboard/src/features/orgs/projects/events/event-triggers/hooks/useDeleteEventTriggerMutation/deleteEventTriggerMigration.ts
@@ -1,5 +1,5 @@
 import type { EventTriggerViewModel } from '@/features/orgs/projects/events/event-triggers/types';
-import { executeMigration } from '@/utils/hasura-api/generated/default/default';
+import { executeMigration } from '@/utils/hasura-api/migrationFetch';
 import type { MigrationOperationOptions } from '@/utils/hasura-api/types';
 
 export interface DeleteEventTriggerMigrationVariables {
