+++ dashboard/src/utils/hasura-api/migrationFetch.test.ts
@@ -1,4 +1,4 @@
-import { executeMigration } from '@/utils/hasura-api/generated/default/default';
+import { executeMigration } from '@/utils/hasura-api/migrationFetch';
 
 const originalEnv = { ...process.env };
 const fetchMock = vi.fn();
